requirements.txt
aiohttp==3.7.2
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3; python_version >= "3.5"
chardet==3.0.4
click==7.1.2
decorator==4.4.2
fsspec==0.8.4
gcsfs==0.7.1
google-api-core==1.23.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-cloud-core==1.4.3
google-cloud-storage==1.32.0
google-crc32c==1.0.0; python_version >= "3.5"
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
idna==2.10
multidict==5.0.2
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.4
pathy==0.3.3
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20; python_version >= "3.5"
python-dateutil==2.8.1
pytz==2020.4
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6; python_version >= "3.5"
six==1.15.0
smart-open==3.0.0
ta==0.6.1
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.25.11
yarl==1.6.3
